refactor: storybook 파일들  디렉토리 변경
