+++ src/stories/Button.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Button from '.';
+import Button from '../components/Button';
 
 export default {
   title: 'Example/Button',
