+++ src/stories/Card.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Card from '.';
+import Card from '../components/Card';
 
 export default {
   title: 'Example/Card',
