+++ src/stories/mixedComponents.stories.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import Button from './Button';
-import Input from './Input';
+import Button from '../components/Button';
+import Input from '../components/Input';
 
 export default {
   title: 'Example/MixedComponents',
